Move action creators setActiveProject -> setActiveExpeirment and deleteProject -> deleteExperiment and delete folder for project action creators
