+++ src/redux/actions/experiments/index.js
@@ -1,15 +1,27 @@
 import createExperiment from './createExperiment';
+import deleteExperiment from './deleteExperiment';
 import updateExperiment from './updateExperiment';
 import loadExperiments from './loadExperiments';
 import loadExperiment from './loadExperiment';
+import setActiveExperiment from './setActiveExperiment';
 import switchExperiment from './switchExperiment';
 import reorderSamples from './reorderSamples';
+import createMetadataTrack from './createMetadataTrack';
+import updateMetadataTrack from './updateMetadataTrack';
+import deleteMetadataTrack from './deleteMetadataTrack';
+import updateValueInMetadataTrack from './updateValueInMetadataTrack';
 
 export {
   createExperiment,
+  deleteExperiment,
   updateExperiment,
   loadExperiments,
   loadExperiment,
+  setActiveExperiment,
   switchExperiment,
   reorderSamples,
+  createMetadataTrack,
+  updateMetadataTrack,
+  deleteMetadataTrack,
+  updateValueInMetadataTrack,
 };
