--- src/redux/actions/projects/index.js
@@ -1,15 +1,0 @@
-import deleteProject from './deleteProject';
-import setActiveProject from './setActiveProject';
-import createMetadataTrack from './createMetadataTrack';
-import updateMetadataTrack from './updateMetadataTrack';
-import deleteMetadataTrack from './deleteMetadataTrack';
-import updateValueInMetadataTrack from './updateValueInMetadataTrack';
-
-export {
-  deleteProject,
-  setActiveProject,
-  createMetadataTrack,
-  updateMetadataTrack,
-  deleteMetadataTrack,
-  updateValueInMetadataTrack,
-};
